(chore) move NewExercise functionality to useNewExercise; extract constants to file; remove unneeded whitespace
